Renamed performance-tests module to jdbc-tests
