+++ appserver/tests/jdbc/src/main/java/org/glassfish/main/test/jdbc/jta/timeout/war/stopwatch/StopwatchInterceptor.java
@@ -14,7 +14,7 @@
  * SPDX-License-Identifier: EPL-2.0 OR GPL-2.0 WITH Classpath-exception-2.0
  */
 
-package org.glassfish.main.test.jta.timeout.war.stopwatch;
+package org.glassfish.main.test.jdbc.jta.timeout.war.stopwatch;
 
 import jakarta.annotation.Priority;
 import jakarta.interceptor.AroundInvoke;
