+++ appserver/tests/jdbc/src/main/resources/jdbc/jta/timeout/war/persistence.xml
