+++ appserver/tests/jdbc/src/test/resources/initSchema.sql
